deli: M-Estimation and Empirical Sandwich Variance Estimation

Solves estimating equations to obtain M-estimators together with empirical sandwich variance estimates, providing a general interface for both custom and built-in estimating equations. Built-in equations cover basic statistics, regression, causal inference, survival analysis, measurement error, and more. Methods follow Stefanski and Boos (2002) doi:10.1198/000313002753631330 and Boos and Stefanski (2013) doi:10.1007/978-1-4614-4818-1_7 . An 'R' port of the 'Python' library 'delicatessen' described in Zivich et al. (2022) doi:10.48550/arXiv.2203.11300 .

See also

Author

Maintainer: Malcolm Barrett malcolmbarrett@gmail.com (ORCID) [copyright holder]

Authors:

Malcolm Barrett malcolmbarrett@gmail.com (ORCID) [copyright holder]

Other contributors:

Paul Zivich (ORCID) (Author of the Python 'delicatessen' library, whose design deli ports) [contributor]